backend/backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
psutil==5.9.6
python-dotenv==1.0.0
PyJWT==2.8.0